+++ CHANGELOG.md
@@ -10,6 +10,7 @@ These are the changes that have not been released yet. The release notes have no
     - Specify kernel module parameters in configuration file.
     - Specify container image build parameters in configuration file.
     - Tag images with module version from configuration file.
+    - Document configuration file in `README.md`.
 
 - Add `--no-prune` option to `build`.